refactor operator data provider modules
